Run the Rust server as a desktop sidecar and fix black-screen playback
- Add RustServerSidecarRuntime: launch library-server as a child process
  with port selection, pairing-token wiring, readiness probing, and
  lifecycle shutdown; wire the shell to it via --rust-sidecar /
  --rust-server-path through the existing remote-client path.
- Add --qa-sidecar-autoplay-first QA hook that fetches the sidecar
  catalog and autoplays the first item for supervised playback QA.
- Fix libmpv resolution in packaged and dev layouts so the native
  executor activates instead of silently falling back to
  command-log-only mode (the original black-screen cause).
- Fix Windows wid formatting to preserve full unsigned 64-bit HWNDs
  instead of truncating to UInt.
- Ignore late diagnostic-log shutdown callbacks after close instead of
  throwing "Stream closed".
- Harden the Rust scanner against unreadable roots/entries (skip
  counters, Windows ACL regression tests) and add an mpv
  request-pattern regression test to the media route.

Verified: cargo fmt/clippy/test (61 tests), :apps:desktop-windows:desktopTest,
and supervised A/B QA proving the same HEVC-10bit episode renders both
from disk and streamed through the Rust sidecar in the packaged app.

+++ apps/desktop-windows/src/desktopMain/kotlin/app/danmaku/desktop/DesktopMpvWindowsOptions.kt
@@ -7,8 +7,8 @@ object DesktopMpvWindowsOptions {
         windowId: Long,
         controlScriptPath: Path? = null,
     ): Map<String, String> {
-        val unsignedWindowId = windowId.toUInt()
-        require(unsignedWindowId != 0u) { "windowId must not be zero" }
+        val unsignedWindowId = windowId.toULong()
+        require(unsignedWindowId != 0uL) { "windowId must not be zero" }
         return embeddedPlaybackOptions(controlScriptPath) + mapOf("wid" to unsignedWindowId.toString())
     }
